Made Homepage -> App and Homepage-componets -> Homepage
